# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TelemetryReading {
  id: number;
  /** @nullable */
  sessionId?: number | null;
  rpm: number;
  tiltX: number;
  tiltY: number;
  tiltZ: number;
  rotationZ?: number;
  avgTiltMag?: number;
  temp?: number;
  throttle: number;
  pwmPct: number;
  vibration?: number;
  autoMode: boolean;
  /** @nullable */
  pidError?: number | null;
  /** @nullable */
  pidOutput?: number | null;
  /** @nullable */
  pidIntegral?: number | null;
  createdAt: string;
}

export interface TelemetryInput {
  /** @nullable */
  sessionId?: number | null;
  rpm: number;
  tiltX: number;
  tiltY: number;
  tiltZ: number;
  rotationZ?: number;
  avgTiltMag?: number;
  temp?: number;
  throttle: number;
  pwmPct: number;
  vibration?: number;
  autoMode: boolean;
  /** @nullable */
  pidError?: number | null;
  /** @nullable */
  pidOutput?: number | null;
  /** @nullable */
  pidIntegral?: number | null;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  running: 'running',
  completed: 'completed',
  error: 'error',
  idle: 'idle',
} as const;

export interface Session {
  id: number;
  name: string;
  status: SessionStatus;
  targetRpm: number;
  duration: number;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
  createdAt: string;
}

export interface SessionInput {
  name: string;
  targetRpm: number;
  duration: number;
}

export type SessionUpdateStatus = typeof SessionUpdateStatus[keyof typeof SessionUpdateStatus];


export const SessionUpdateStatus = {
  running: 'running',
  completed: 'completed',
  error: 'error',
  idle: 'idle',
} as const;

export interface SessionUpdate {
  status?: SessionUpdateStatus;
  endedAt?: string;
}

export type CommandInputCommand = typeof CommandInputCommand[keyof typeof CommandInputCommand];


export const CommandInputCommand = {
  MOTOR_START: 'MOTOR_START',
  MOTOR_STOP: 'MOTOR_STOP',
  EMERGENCY_STOP: 'EMERGENCY_STOP',
  SPEED_UP: 'SPEED_UP',
  SPEED_DOWN: 'SPEED_DOWN',
  RESET_REF: 'RESET_REF',
  AUTO_ON: 'AUTO_ON',
  AUTO_OFF: 'AUTO_OFF',
} as const;

export type CommandInputKey = typeof CommandInputKey[keyof typeof CommandInputKey];


export const CommandInputKey = {
  S: 'S',
  X: 'X',
  H: 'H',
  L: 'L',
  Z: 'Z',
  A: 'A',
} as const;

export interface CommandInput {
  command: CommandInputCommand;
  key?: CommandInputKey;
}

export interface CommandResult {
  success: boolean;
  message: string;
}

export type ListTelemetryParams = {
limit?: number;
/**
 * @nullable
 */
sessionId?: number | null;
};