# lib/api-zod/src/generated/types/commandInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { CommandInputCommand } from './commandInputCommand';
import type { CommandInputKey } from './commandInputKey';

export interface CommandInput {
  command: CommandInputCommand;
  key?: CommandInputKey;
}