# lib/api-zod/src/generated/types/commandResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface CommandResult {
  success: boolean;
  message: string;
}