# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatus } from './sessionStatus';

export interface Session {
  id: number;
  name: string;
  status: SessionStatus;
  targetRpm: number;
  duration: number;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  endedAt?: Date | null;
  createdAt: Date;
}