# lib/api-zod/src/generated/types/sessionUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */
import type { SessionUpdateStatus } from './sessionUpdateStatus';

export interface SessionUpdate {
  status?: SessionUpdateStatus;
  endedAt?: Date;
}