# lib/api-zod/src/generated/types/telemetryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gyro Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface TelemetryInput {
  /** @nullable */
  sessionId?: number | null;
  rpm: number;
  tiltX: number;
  tiltY: number;
  tiltZ: number;
  rotationZ?: number;
  avgTiltMag?: number;
  temp?: number;
  throttle: number;
  pwmPct: number;
  vibration?: number;
  autoMode: boolean;
  /** @nullable */
  pidError?: number | null;
  /** @nullable */
  pidOutput?: number | null;
  /** @nullable */
  pidIntegral?: number | null;
}